+++ src/design/plone/contenttypes/events/common.py
@@ -121,8 +121,8 @@ SUBFOLDERS_MAPPING = {
         {"id": "altre-cariche", "title": "Altre cariche", "allowed_types": ("File",)},
         {"id": "incarichi", "title": "Incarichi", "allowed_types": ("Incarico",)},
         {
-            "id": "altri-allegati",
-            "title": "Altri Allegati",
+            "id": "altri-documenti",
+            "title": "Altri Documenti",
             "allowed_types": ("File", "Image", "Link"),
         },
     ],