+++ src/design/plone/contenttypes/restapi/deserializers/venue.py
@@ -1,9 +1,8 @@
 # -*- coding: utf-8 -*-
 from collective.venue.interfaces import IVenue
-from plone.restapi.behaviors import IBlocks
+from design.plone.contenttypes.utils import text_in_block
 from plone.restapi.deserializer import json_body
 from plone.restapi.deserializer.dxcontent import DeserializeFromJson
-from plone.restapi.indexers import SearchableText_blocks
 from plone.restapi.interfaces import IDeserializeFromJson
 from zExceptions import BadRequest
 from zope.component import adapter
@@ -19,30 +18,6 @@ MANDATORY_RICH_TEXT_FIELDS = ["modalita_accesso"]
 
 def new_error(message):
     return {"error": "ValidationError", "message": message}
-
-
-def text_in_block(blocks):
-    @implementer(IBlocks)
-    class FakeObject(object):
-        """
-        We use a fake object to use SearchableText Indexer
-        """
-
-        def Subject(self):
-            return ""
-
-        def __init__(self, blocks, blocks_layout):
-            self.blocks = blocks
-            self.blocks_layout = blocks_layout
-            self.id = ""
-            self.title = ""
-            self.description = ""
-
-    if not blocks:
-        return None
-
-    fakeObj = FakeObject(blocks.get("blocks", ""), blocks.get("blocks_layout", ""))
-    return SearchableText_blocks(fakeObj)()
 
 
 @implementer(IDeserializeFromJson)